Update breakpoints for "tablet" and "netbook" media queries
Now that everything is mobile first, it was time to revisit these
breakpoints. 768px was only chosen because of the dimensions of the
iPad, and using that exact number in a media query meant that the iPad
would still be displaying the desktop/netbook media queries. Instead,
we’re rounding up to 800px, which is a more generic breakpoint (and a
nicer number).